Move all streams to new Streams module and replace ProxyFileStream with GrowingFileStream.
Instead of ProxyFileStream, now use a combination of GrowingFileStream
and StreamToFile like the PeerManager does to download and upload. It's
mostly working, only some errors when aborting apt-get in test 'a'.

Also added a new test 'c' for downloading out of order, which is why
ProxyFileStream can not be used.